requirements.txt
beautifulsoup4==4.13.5
cloudscraper==1.2.71
curl_cffi==0.12.0
Flask==3.1.2
keyboard==0.13.5
orjson==3.11.3
pywebview==5.5.dev16+g21a609b
Requests==2.32.5
websocket_client==1.8.0
websockets==15.0
